requirements.txt
asgiref==3.2.3
Django==3.0.2
django-ckeditor==5.9.0
django-filter==2.2.0
django-js-asset==1.2.2
gunicorn==20.0.4
Pillow==7.0.0
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
